Extract CountriesMapData
